Rename file to program
